Convert int to bigInt in the concerned method

+++ src/main/java/com/iexec/common/chain/ChainTask.java
@@ -139,6 +139,6 @@ public class ChainTask {
     }
 
     public String getChainTaskId() {
-        return ChainUtils.generateChainTaskId(dealid, BigInteger.valueOf(idx));
+        return ChainUtils.generateChainTaskId(dealid, idx);
     }
 }
